Prepare application for testing hooks
 - setup proxy
 - setup tests
 - setup thunk middleware
 - setup store, actions and apis
 - add root component
 - add Xkcd component that uses api, state, reducer, action

That Xkcd component will be used to keep other components in check when
they will have it included in them. But when mount rendering will be
used, it will throw as store for Xkcd will not be in context.

And that is what all of the tests here will be about - how to use
all of the hooks (if possible) without having to mount in a way that
dependencies for child components will be needed.